+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - Improve word selection and related editing actions to treat commas as word delimiters.
 - [trivial] Improve stability when handling file change conflicts.
-- [beta] Update the outline extraction for C#.
+- [beta] Update outline extraction algorithms for C#, Java, JavaScript, PHP, and TypeScript.
 - [beta][trivial] Update the document icon for the CotEditor Syntax document.